move functions to azureFunctionsUtils
